requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-countries==6.1.2
django-dob-widget==1.1.3
django-easy-timezones==0.8.0
django-ip==1.0.2
django-phone-field==1.8.0
django-six==1.0.4
geoip2==3.0.0
idna==2.9
ipaddress==1.0.16
maxminddb==1.5.3
oauthlib==3.1.0
Pillow==7.1.2
pygeoip==0.3.2
python3-openid==3.1.0
pytz==2016.6
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
